IC.Canister.StableMemory

Description

This module provides a wrapper around primitive byte array, exposing just the bits needed for accessing the stable memory.

Documentation

type Memory s = STRef s Repr

data Repr

Immutable representation of stable memory that can be used for snapshotting.

new :: HostM s (Memory s)

Constructs a new empty memory.

size :: Memory s -> HostM s Size

Returns the size of stable memory in WebAssembly pages.

grow :: Memory s -> Size -> HostM s Size

Attempts to grow stable memory by delta pages.

read :: Memory s -> Address -> Size -> HostM s ByteString

Reads a range of bytes from memory.

write :: Memory s -> Address -> ByteString -> HostM s ()

Writes a byte string at the specified offset.

export :: Memory s -> ST s Repr

Exports immutable memory representation.

imp :: Memory s -> Repr -> ST s ()

Sets the contents of memory to a previously exported value.

serialize :: Repr -> ByteString

Converts internal memory representation into a bytestring.

deserialize :: ByteString -> Repr

Constructs internal memory representation from a blob. Throws an exception if the length of the blob is not a multiple of 64KiB.