IC.Ref.Types

Description

This module implements the main abstract types of the Internet Computer.

Documentation

data CallRequest

Constructors

CallRequest CanisterId UserId MethodName Blob

data QueryRequest

Constructors

QueryRequest CanisterId UserId MethodName Blob

data ReadStateRequest

Constructors

ReadStateRequest UserId [Path]

data RequestStatus

Constructors

Received

Processing

CallResponse CallResponse

data CallResponse

Constructors

Rejected (RejectCode, String, Maybe ErrorCode)

Replied Blob

data ValidationError

Constructors

HTTPError String

ExecutionError (RejectCode, String, Maybe ErrorCode)

data ReqResponse

Constructors

RequestError (RejectCode, String, Maybe ErrorCode)

QueryResponse CallResponse

ReadStateResponse Certificate

data RunStatus

Constructors

IsRunning

IsStopping [CallId]

IsStopped

IsDeleted

data CanisterContent

Constructors

CanisterContent

Fields

can_mod :: CanisterModule

wasm_state :: WasmState

data CanisterHistory

Constructors

CanisterHistory

Fields

changes :: [Change]

total_num_changes :: Word64

data CanState

Constructors

CanState

Fields

content :: Maybe CanisterContent

run_status :: RunStatus

controllers :: Set EntityId

memory_allocation :: Natural

compute_allocation :: Natural

freezing_threshold :: Natural

time :: Timestamp

cycle_balance :: Natural

certified_data :: Blob

canister_version :: Natural

canister_history :: CanisterHistory

global_timer :: Natural

last_action :: Maybe EntryPoint

Not part of the spec, but in this implementation we schedule heartbeats only for canisters who have not been idle since the last heartbeat, so we remember the last action.

data EntryPoint

Constructors

Public MethodName Blob

Closure Callback Response Cycles

Heartbeat

GlobalTimer

type CallId = Int

data CallContext

Constructors

CallContext

Fields

canister :: CanisterId

origin :: CallOrigin

needs_to_respond :: NeedsToRespond

deleted :: Bool

available_cycles :: Cycles

last_trap :: Maybe String

non-normative, but yields better reject messages

data CallOrigin

Constructors

FromUser RequestID CanisterId

FromCanister CallId Callback

FromSystemTask

data Message

Constructors

CallMessage

Fields

call_context :: CallId

entry :: EntryPoint

ResponseMessage

Fields

call_context :: CallId

response :: Response

refunded_cycles :: Cycles

type Subnet = (EntityId, SubnetType, Word64, SecretKey, [(Word64, Word64)])

isRootSubnet :: Subnet -> Bool

data IC

Constructors

IC

Fields

canisters :: CanisterId ↦ CanState

requests :: RequestID ↦ (CallRequest, (RequestStatus, CanisterId))

messages :: Seq Message

call_contexts :: CallId ↦ CallContext

rng :: StdGen

secretRootKey :: SecretKey

rootSubnet :: Maybe EntityId

subnets :: [Subnet]

type ICM m = (MonadState IC m, HasRefConfig, HasCallStack, MonadIO m)

initialIC :: [SubnetConfig] -> IO IC

type CanReject = MonadError (RejectCode, String, Maybe ErrorCode)

reject :: CanReject m => RejectCode -> String -> Maybe ErrorCode -> m a2

canisterRejected :: (RejectCode, String, Maybe ErrorCode) -> CallResponse

rejectMessage :: (RejectCode, String, Maybe ErrorCode) -> String

onReject :: ICM m => ((RejectCode, String, Maybe ErrorCode) -> m b) -> (forall m'. (CanReject m', ICM m') => m' b) -> m b

onErr :: Monad m => m (Either a b) -> (a -> m b) -> m b

orElse :: Monad m => m (Maybe a) -> m a -> m a

onTrap :: Monad m => m (TrapOr a) -> (String -> m a) -> m a

getSubnetFromCanisterId' :: ICM m => CanisterId -> m (Maybe Subnet)

getSubnetFromCanisterId :: (CanReject m, ICM m) => CanisterId -> m Subnet

getSubnetFromSubnetId :: (CanReject m, ICM m) => CanisterId -> m (Maybe Subnet)

getCanisterRootKey :: CanisterId -> ExtendedSecretKey

canisterMustExist :: (CanReject m, ICM m) => CanisterId -> m ()

isCanisterEmpty :: ICM m => CanisterId -> m Bool

createEmptyCanister :: ICM m => CanisterId -> Set EntityId -> Timestamp -> m ()

enqueueMessage :: ICM m => Message -> m ()

managementCanisterId :: EntityId

callerOfCallRequest :: CallRequest -> EntityId

callerOfRequest :: ICM m => RequestID -> m EntityId

callerOfCallID :: ICM m => CallId -> m EntityId

calleeOfCallID :: ICM m => CallId -> m EntityId

ecidOfCallID :: ICM m => CallId -> m CanisterId

getCallContext :: ICM m => CallId -> m CallContext

modifyCallContext :: ICM m => CallId -> (CallContext -> CallContext) -> m ()

getCallContextCycles :: ICM m => CallId -> m Cycles

setCallContextCycles :: ICM m => CallId -> Cycles -> m ()

respondCallContext :: ICM m => CallId -> Response -> m ()

replyCallContext :: ICM m => CallId -> Blob -> m ()

rejectCallContext :: ICM m => CallId -> (RejectCode, String, Maybe ErrorCode) -> m ()

deleteCallContext :: ICM m => CallId -> m ()

getCanister :: ICM m => CanisterId -> m CanState

modCanister :: ICM m => CanisterId -> (CanState -> CanState) -> m ()

setCanisterContent :: ICM m => CanisterId -> CanisterContent -> m ()

modCanisterContent :: ICM m => CanisterId -> (CanisterContent -> CanisterContent) -> m ()

setCanisterState :: ICM m => CanisterId -> WasmState -> m ()

getControllers :: ICM m => CanisterId -> m (Set EntityId)

setControllers :: ICM m => CanisterId -> Set EntityId -> m ()

setComputeAllocation :: ICM m => CanisterId -> Natural -> m ()

setMemoryAllocation :: ICM m => CanisterId -> Natural -> m ()

setFreezingThreshold :: ICM m => CanisterId -> Natural -> m ()

getBalance :: ICM m => CanisterId -> m Natural

setBalance :: ICM m => CanisterId -> Natural -> m ()

setCertifiedData :: ICM m => CanisterId -> Blob -> m ()

getRunStatus :: ICM m => CanisterId -> m RunStatus

setRunStatus :: ICM m => CanisterId -> RunStatus -> m ()

getCanisterState :: ICM m => CanisterId -> m WasmState

getCanisterMod :: ICM m => CanisterId -> m CanisterModule

getCanisterTime :: ICM m => CanisterId -> m Timestamp

getCanisterVersion :: ICM m => CanisterId -> m Natural

bumpCanisterVersion :: ICM m => CanisterId -> m ()

getCanisterHistory :: ICM m => CanisterId -> m CanisterHistory

addCanisterHistory :: ICM m => CanisterId -> Change -> m ()

getCanisterGlobalTimer :: ICM m => CanisterId -> m Natural

setCanisterGlobalTimer :: ICM m => CanisterId -> Natural -> m ()

module_hash :: CanState -> Maybe Blob

idle_cycles_burned_per_day :: CanState -> Natural

canisterEnv :: ICM m => CanisterId -> m Env

performCanisterActions :: ICM m => CanisterId -> CanisterActions -> m ()