IC.StateFile

Description

This module provides an (optionally) persistent store for the state of the IC.

It provides: * atomic, blocking, exclusive write access * read access to the latest state, at any time

If not backed by a file, it is simply kept in memory.

If backed by a file, state changes are serialized by a separate thread. It waits for a state change to happen, then waits for 10s, and then serializes the then youngest state. If the process is shut down (via conclude), it serializes immediately.

Serialization is atomic (write to a temporary file, then move), so it _should_ be safe to kill the process; _a_ recent state will be persisted.

Documentation

data Store a

withStore :: Serialise a => IO a -> Maybe FilePath -> (Store a -> IO b) -> IO b

modifyStore :: Serialise a => Store a -> StateT a IO b -> IO b

peekStore :: Store a -> IO a