IC.Test.Agent

Description

This module contains function to interact with an Internet Computer instance.

The primary customer here is IC.Test.Spec, i.e the test suite. Therefore, the functions here give access to varoius levels of abstractions, and gives more control than a “normal” agent library would do.

Also, because of the focus on testing, failures are repoted directly with HUnit’s assertFailure.

As guidance: This modules does _not_ rely on the universal canister.

This module can also be used in a REPL; see connect.

Documentation

type HTTPErrOr a = Either (Int, String) a

type HasAgentConfig = ?agentConfig :: AgentConfig

type IC00 = Blob -> Text -> Blob -> IO ReqResponse

data ReqResponse

Constructors

Reply Blob

Reject Natural Text (Maybe Text)

data ReqStatus

Constructors

Processing

Pending

Responded ReqResponse

UnknownStatus

data AgentConfig

Constructors

AgentConfig

Fields

tc_root_key :: Blob

tc_manager :: Manager

tc_endPoint :: String

tc_test_port :: Int

tc_timeout :: Int

addExpiry :: GenR -> IO GenR

addNonce :: GenR -> IO GenR

addNonceExpiryEnv :: GenR -> IO GenR

anonymousUser :: Blob

as2Word64 :: HasCallStack => Blob -> IO (Word64, Word64)

asHex :: Blob -> String

asRight :: HasCallStack => Either Text a -> IO a

asWord128 :: HasCallStack => Blob -> IO Natural

asWord32 :: HasCallStack => Blob -> IO Word32

asWord64 :: HasCallStack => Blob -> IO Word64

awaitCall' :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (HTTPErrOr ReqResponse)

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit". Returns either a HTTP Error code, or if the status is 2xx, poll for the request response, and return decoded CBOR

awaitCall :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO ReqResponse

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit", poll for the request response, and return decoded CBOR

awaitKnown :: HasAgentConfig => IO ReqStatus -> IO ReqStatus

awaitStatus :: HasAgentConfig => IO ReqStatus -> IO ReqResponse

bothSame :: (Eq a, Show a) => (a, a) -> Assertion

certValue :: HasCallStack => CertVal a => Certificate -> [Blob] -> IO a

certValueAbsent :: HasCallStack => Certificate -> [Blob] -> IO ()

code202 :: HasCallStack => Response ByteString -> IO ()

code202_or_4xx :: HasCallStack => Response ByteString -> IO ()

code2xx :: HasCallStack => Response ByteString -> IO ()

code4xx :: HasCallStack => Response ByteString -> IO ()

connect :: String -> Int -> Int -> IO ReplWrapper

This is for use from the Haskell REPL, see README.md

decodeCert' :: HasCallStack => Blob -> IO Certificate

defaultSK :: SecretKey

defaultUser :: Blob

delegationEnv :: SecretKey -> [(SecretKey, Maybe [Blob])] -> GenR -> IO GenR

doesn'tExist :: Blob

ecdsaSK :: SecretKey

ecdsaUser :: Blob

enum :: (AllUniqueLabels r, KnownSymbol l, (r .! l) ~ ()) => Label l -> Var r

envelope :: SecretKey -> GenR -> IO GenR

envelopeFor :: Blob -> GenR -> IO GenR

extractCertData :: Blob -> Blob -> IO Blob

getRequestStatus :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Blob -> IO ReqStatus

getStateCert' :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO (Response Blob)

getStateCert :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO Certificate

ic00 :: HasAgentConfig => IC00

ic00as :: (HasAgentConfig, HasCallStack) => Blob -> IC00

ingressDelay :: IO ()

is2xx :: HasCallStack => HTTPErrOr a -> IO a

isErrOrReject :: HasCallStack => [Natural] -> HTTPErrOr ReqResponse -> IO ()

isPendingOrProcessing :: ReqStatus -> IO ()

isReject :: HasCallStack => [Natural] -> ReqResponse -> IO ()

isReply :: HasCallStack => ReqResponse -> IO Blob

okCBOR :: HasCallStack => Response ByteString -> IO GenR

otherSK :: SecretKey

otherUser :: Blob

makeAgentConfig :: String -> Int -> Int -> IO AgentConfig

postCBOR :: (HasCallStack, HasAgentConfig) => String -> GenR -> IO (Response ByteString)

Posting a CBOR request, returning a raw bytestring

postCallCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

postQueryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

postReadStateCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

preFlight :: OptionSet -> IO AgentConfig

queryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO GenR

Add envelope to CBOR request, add a nonce and expiry if it is not there, post to "read", return decoded CBOR

queryResponse :: GenR -> IO ReqResponse

runGet :: HasCallStack => Get a -> Blob -> IO a

secp256k1SK :: SecretKey

secp256k1User :: Blob

senderOf :: GenR -> Blob

shorten :: Int -> String -> String

submitCall :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (IO ReqStatus)

textual :: Blob -> String

validateStateCert :: (HasCallStack, HasAgentConfig) => Certificate -> IO ()

verifySignature :: Blob -> Blob -> Blob -> Bool

webAuthnECDSASK :: SecretKey

webAuthnECDSAUser :: Blob

webAuthnRSASK :: SecretKey

webAuthnRSAUser :: Blob

withAgentConfig :: (forall. HasAgentConfig => a) -> AgentConfig -> a

callIC :: forall s a b. (HasCallStack, HasAgentConfig) => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => (CandidArg a, CandidArg b) => IC00 -> Blob -> Label s -> a -> IO b

callIC' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => IC00 -> Blob -> Label s -> a -> IO ReqResponse

callIC'' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => Blob -> Blob -> Label s -> a -> IO (HTTPErrOr ReqResponse)

agentConfig :: HasAgentConfig => AgentConfig