IC.Test.Universal

Description

Helpers to program the “Universal module”. This is essentially a small, type-safe DSL to produce the small stack-based programming language interpreted by the universal canister.

This DSL is expression-based, not stack based; seems to suite all our needs and is simpler to work with.

This language is not stable; therefore there is no separarte documentation of specification than this file and `universal-canistersrc`

Documentation

data T

Constructors

I

I64

B

newtype Exp (result :: T) where

Constructors

Exp :: Builder -> Exp a

newtype Prog where

Constructors

Prog :: Builder -> Prog

run :: Prog -> ByteString

(>>>) :: Prog -> Prog -> Prog

class Op a where

Methods

mkOp :: Word8 -> Builder -> a

op :: Op a => Word8 -> a

noop :: Prog

ignore :: Exp t -> Prog

int :: Word32 -> Exp 'I

int64 :: Word64 -> Exp 'I64

bytes :: ByteString -> Exp 'B

replyDataAppend :: Exp 'B -> Prog

reply :: Prog

self :: Exp 'B

reject :: Exp 'B -> Prog

caller :: Exp 'B

reject_msg :: Exp 'B

reject_code :: Exp 'I

i2b :: Exp 'I -> Exp 'B

i64tob :: Exp 'I64 -> Exp 'B

argData :: Exp 'B

cat :: Exp 'B -> Exp 'B -> Exp 'B

stableSize :: Exp 'I

stableGrow :: Exp 'I -> Exp 'I

stableRead :: Exp 'I -> Exp 'I -> Exp 'B

stableWrite :: Exp 'I -> Exp 'B -> Prog

getTime :: Exp 'I64

getAvailableCycles :: Exp 'I64

getBalance :: Exp 'I64

getRefund :: Exp 'I64

acceptCycles :: Exp 'I64 -> Exp 'I64

debugPrint :: Exp 'B -> Prog

trap :: Exp 'B -> Prog

setGlobal :: Exp 'B -> Prog

getGlobal :: Exp 'B

badPrint :: Prog

onPreUpgrade :: Exp 'B -> Prog

callNew :: Exp 'B -> Exp 'B -> Exp 'B -> Exp 'B -> Prog

callDataAppend :: Exp 'B -> Prog

callCyclesAdd :: Exp 'I64 -> Prog

callPerform :: Prog

setCertifiedData :: Exp 'B -> Prog

getCertificatePresent :: Exp 'I

getCertificate :: Exp 'B

getStatus :: Exp 'I

acceptMessage :: Prog

onInspectMessage :: Exp 'B -> Prog

methodName :: Exp 'B

trapIfEq :: Exp 'B -> Exp 'B -> Exp 'B -> Prog

callOnCleanup :: Exp 'B -> Prog

stable64Size :: Exp 'I64

stable64Grow :: Exp 'I64 -> Exp 'I64

stable64Read :: Exp 'I64 -> Exp 'I64 -> Exp 'B

stable64Write :: Exp 'I64 -> Exp 'B -> Prog

onHeartbeat :: Exp 'B -> Prog

performanceCounter :: Exp 'I -> Exp 'I64

getBalance128 :: Exp 'B

getAvailableCycles128 :: Exp 'B

getRefund128 :: Exp 'B

acceptCycles128 :: Exp 'I64 -> Exp 'I64 -> Exp 'B

callCyclesAdd128 :: Exp 'I64 -> Exp 'I64 -> Prog

onTransform :: Exp 'B -> Prog

callback :: Prog -> Exp 'B

replyData :: Exp 'B -> Prog

data CallArgs

Constructors

CallArgs

Fields

on_reply :: Prog

on_reject :: Prog

on_cleanup :: Maybe Prog

other_side :: Prog

cycles :: Word64

icpts :: Word64

inter_call :: ByteString -> ByteString -> CallArgs -> Prog

inter_update :: ByteString -> CallArgs -> Prog

inter_query :: ByteString -> CallArgs -> Prog

defArgs :: CallArgs

By default, the other side responds with some text indicating caller and callee, and the callbacks reply with the response.

defaultOtherSide :: Prog

relayReply :: Prog

relayReject :: Prog