IC.Wasm.Winter.Persist

Description

This module provides a way to persist the state of a Winter Wasm instance, and to recover it.

It is tailored to the use by ic-ref. For example it assumes that the table of a wasm instance is immutable.

Documentation

newtype PInstance

This stores data read from an instance.

Constructors

PInstance (Persisted (Instance ()))

data PModuleInst

Constructors

PModuleInst

Fields

memories :: Vector (Persisted (MemoryInst (ST ())))

globals :: Vector (Persisted (GlobalInst (ST ())))

persistInstance :: Instance s -> ST s PInstance

resumeInstance :: Instance s -> PInstance -> ST s ()

persistMemory :: MemoryInst (ST s) -> ST s ByteString

resumeMemory :: MemoryInst (ST s) -> ByteString -> ST s ()