IC.Wasm.WinterMemory

Description

This module provides a thin wrapper around the winter Wasm engine, exposing just the bits needed for accessing the stable memory. This will hopefully go away once multiple memories are supported.

Documentation

type Memory s = MemoryInst (ST s)

new :: HostM s (Memory s)

size :: Memory s -> HostM s Size

grow :: Memory s -> Size -> HostM s Size

read :: Memory s -> Address -> Size -> HostM s ByteString

write :: Memory s -> Address -> ByteString -> HostM s ()

export :: Memory s -> ST s ByteString

imp :: Memory s -> ByteString -> ST s ()