SourceId

Description

This module exports a source id in a way that works well within a nix build (no .git available) and outside nix, according to this logic:

If git works, use `git describe`

Else, if $out is set (so this is a nix build), extract an identifer from the out hash

Else, it says something like unidentified build.

This is an early experiment. If successful, this logic ought to move into a library, and maybe also implemented for rust artifacts. (see RPL-101)

Note that cabal build will not recompile this module if it does not have to, so in local development, this is less reliable than it should. See https:/www.joachim-breitner.deblog/772-Template_Haskell_recompilation for more details.

Documentation

id :: String