IC.Debug.JSON

Description

This module defines ToJSON instances of the IC state.

We put them into their own module, despite the usual advise against orphan instances, to emphasize that these are there just for debugging purposes using `ic-ref`.

(Why JSON? Because Browsers render them nicely in a interactive display where you can open and collapse subcomponents – much easier to get this feature this way, compared to writing custom HTML output.)

Orphan instances

ToJSON ByteString

Instance details

Methods

toJSON :: ByteString -> Value

toEncoding :: ByteString -> Encoding

toJSONList :: [ByteString] -> Value

toEncodingList :: [ByteString] -> Encoding

ToJSON CanisterModule

Instance details

Methods

toJSON :: CanisterModule -> Value

toEncoding :: CanisterModule -> Encoding

toJSONList :: [CanisterModule] -> Value

toEncodingList :: [CanisterModule] -> Encoding

ToJSON CanisterSnapshot

Instance details

Methods

toJSON :: CanisterSnapshot -> Value

toEncoding :: CanisterSnapshot -> Encoding

toJSONList :: [CanisterSnapshot] -> Value

toEncodingList :: [CanisterSnapshot] -> Encoding

ToJSON Repr

Instance details

Methods

toJSON :: Repr -> Value

toEncoding :: Repr -> Encoding

toJSONList :: [Repr] -> Value

toEncodingList :: [Repr] -> Encoding

ToJSON SecretKey

Instance details

Methods

toJSON :: SecretKey -> Value

toEncoding :: SecretKey -> Encoding

toJSONList :: [SecretKey] -> Value

toEncodingList :: [SecretKey] -> Encoding

ToJSON CallContext

Instance details

Methods

toJSON :: CallContext -> Value

toEncoding :: CallContext -> Encoding

toJSONList :: [CallContext] -> Value

toEncodingList :: [CallContext] -> Encoding

ToJSON CallOrigin

Instance details

Methods

toJSON :: CallOrigin -> Value

toEncoding :: CallOrigin -> Encoding

toJSONList :: [CallOrigin] -> Value

toEncodingList :: [CallOrigin] -> Encoding

ToJSON CallRequest

Instance details

Methods

toJSON :: CallRequest -> Value

toEncoding :: CallRequest -> Encoding

toJSONList :: [CallRequest] -> Value

toEncodingList :: [CallRequest] -> Encoding

ToJSON CallResponse

Instance details

Methods

toJSON :: CallResponse -> Value

toEncoding :: CallResponse -> Encoding

toJSONList :: [CallResponse] -> Value

toEncodingList :: [CallResponse] -> Encoding

ToJSON CanState

Instance details

Methods

toJSON :: CanState -> Value

toEncoding :: CanState -> Encoding

toJSONList :: [CanState] -> Value

toEncodingList :: [CanState] -> Encoding

ToJSON CanisterContent

Instance details

Methods

toJSON :: CanisterContent -> Value

toEncoding :: CanisterContent -> Encoding

toJSONList :: [CanisterContent] -> Value

toEncodingList :: [CanisterContent] -> Encoding

ToJSON CanisterHistory

Instance details

Methods

toJSON :: CanisterHistory -> Value

toEncoding :: CanisterHistory -> Encoding

toJSONList :: [CanisterHistory] -> Value

toEncodingList :: [CanisterHistory] -> Encoding

ToJSON EntryPoint

Instance details

Methods

toJSON :: EntryPoint -> Value

toEncoding :: EntryPoint -> Encoding

toJSONList :: [EntryPoint] -> Value

toEncodingList :: [EntryPoint] -> Encoding

ToJSON IC

Instance details

Methods

toJSON :: IC -> Value

toEncoding :: IC -> Encoding

toJSONList :: [IC] -> Value

toEncodingList :: [IC] -> Encoding

ToJSON Message

Instance details

Methods

toJSON :: Message -> Value

toEncoding :: Message -> Encoding

toJSONList :: [Message] -> Value

toEncodingList :: [Message] -> Encoding

ToJSON RequestStatus

Instance details

Methods

toJSON :: RequestStatus -> Value

toEncoding :: RequestStatus -> Encoding

toJSONList :: [RequestStatus] -> Value

toEncodingList :: [RequestStatus] -> Encoding

ToJSON RunStatus

Instance details

Methods

toJSON :: RunStatus -> Value

toEncoding :: RunStatus -> Encoding

toJSONList :: [RunStatus] -> Value

toEncodingList :: [RunStatus] -> Encoding

ToJSON Callback

Instance details

Methods

toJSON :: Callback -> Value

toEncoding :: Callback -> Encoding

toJSONList :: [Callback] -> Value

toEncodingList :: [Callback] -> Encoding

ToJSON CanisterInstallMode

Instance details

Methods

toJSON :: CanisterInstallMode -> Value

toEncoding :: CanisterInstallMode -> Encoding

toJSONList :: [CanisterInstallMode] -> Value

toEncodingList :: [CanisterInstallMode] -> Encoding

ToJSON Change

Instance details

Methods

toJSON :: Change -> Value

toEncoding :: Change -> Encoding

toJSONList :: [Change] -> Value

toEncodingList :: [Change] -> Encoding

ToJSON ChangeDetails

Instance details

Methods

toJSON :: ChangeDetails -> Value

toEncoding :: ChangeDetails -> Encoding

toJSONList :: [ChangeDetails] -> Value

toEncodingList :: [ChangeDetails] -> Encoding

ToJSON ChangeOrigin

Instance details

Methods

toJSON :: ChangeOrigin -> Value

toEncoding :: ChangeOrigin -> Encoding

toJSONList :: [ChangeOrigin] -> Value

toEncodingList :: [ChangeOrigin] -> Encoding

ToJSON EntityId

Instance details

Methods

toJSON :: EntityId -> Value

toEncoding :: EntityId -> Encoding

toJSONList :: [EntityId] -> Value

toEncodingList :: [EntityId] -> Encoding

ToJSON ErrorCode

Instance details

Methods

toJSON :: ErrorCode -> Value

toEncoding :: ErrorCode -> Encoding

toJSONList :: [ErrorCode] -> Value

toEncodingList :: [ErrorCode] -> Encoding

ToJSON MethodCall

Instance details

Methods

toJSON :: MethodCall -> Value

toEncoding :: MethodCall -> Encoding

toJSONList :: [MethodCall] -> Value

toEncodingList :: [MethodCall] -> Encoding

ToJSON NeedsToRespond

Instance details

Methods

toJSON :: NeedsToRespond -> Value

toEncoding :: NeedsToRespond -> Encoding

toJSONList :: [NeedsToRespond] -> Value

toEncodingList :: [NeedsToRespond] -> Encoding

ToJSON RejectCode

Instance details

Methods

toJSON :: RejectCode -> Value

toEncoding :: RejectCode -> Encoding

toJSONList :: [RejectCode] -> Value

toEncodingList :: [RejectCode] -> Encoding

ToJSON Response

Instance details

Methods

toJSON :: Response -> Value

toEncoding :: Response -> Encoding

toJSONList :: [Response] -> Value

toEncodingList :: [Response] -> Encoding

ToJSON SubnetType

Instance details

Methods

toJSON :: SubnetType -> Value

toEncoding :: SubnetType -> Encoding

toJSONList :: [SubnetType] -> Value

toEncodingList :: [SubnetType] -> Encoding

ToJSON Timestamp

Instance details

Methods

toJSON :: Timestamp -> Value

toEncoding :: Timestamp -> Encoding

toJSONList :: [Timestamp] -> Value

toEncodingList :: [Timestamp] -> Encoding

ToJSON WasmClosure

Instance details

Methods

toJSON :: WasmClosure -> Value

toEncoding :: WasmClosure -> Encoding

toJSONList :: [WasmClosure] -> Value

toEncodingList :: [WasmClosure] -> Encoding

ToJSON PInstance

Instance details

Methods

toJSON :: PInstance -> Value

toEncoding :: PInstance -> Encoding

toJSONList :: [PInstance] -> Value

toEncodingList :: [PInstance] -> Encoding

ToJSON PModuleInst

Instance details

Methods

toJSON :: PModuleInst -> Value

toEncoding :: PModuleInst -> Encoding

toJSONList :: [PModuleInst] -> Value

toEncodingList :: [PModuleInst] -> Encoding

ToJSON StdGen

Instance details

Methods

toJSON :: StdGen -> Value

toEncoding :: StdGen -> Encoding

toJSONList :: [StdGen] -> Value

toEncodingList :: [StdGen] -> Encoding

ToJSON Value

Instance details

Methods

toJSON :: Value -> Value0

toEncoding :: Value -> Encoding

toJSONList :: [Value] -> Value0

toEncodingList :: [Value] -> Encoding

ToJSONKey ByteString

Instance details

Methods

toJSONKey :: ToJSONKeyFunction ByteString

toJSONKeyList :: ToJSONKeyFunction [ByteString]

ToJSONKey EntityId

Instance details

Methods

toJSONKey :: ToJSONKeyFunction EntityId

toJSONKeyList :: ToJSONKeyFunction [EntityId]

Generic CanisterSnapshot

Instance details

Associated Types

type Rep CanisterSnapshot :: Type -> Type

Methods

from :: CanisterSnapshot -> Rep CanisterSnapshot x

to :: Rep CanisterSnapshot x -> CanisterSnapshot

Generic CallContext

Instance details

Associated Types

type Rep CallContext :: Type -> Type

Methods

from :: CallContext -> Rep CallContext x

to :: Rep CallContext x -> CallContext

Generic CallOrigin

Instance details

Associated Types

type Rep CallOrigin :: Type -> Type

Methods

from :: CallOrigin -> Rep CallOrigin x

to :: Rep CallOrigin x -> CallOrigin

Generic CallRequest

Instance details

Associated Types

type Rep CallRequest :: Type -> Type

Methods

from :: CallRequest -> Rep CallRequest x

to :: Rep CallRequest x -> CallRequest

Generic CallResponse

Instance details

Associated Types

type Rep CallResponse :: Type -> Type

Methods

from :: CallResponse -> Rep CallResponse x

to :: Rep CallResponse x -> CallResponse

Generic CanState

Instance details

Associated Types

type Rep CanState :: Type -> Type

Methods

from :: CanState -> Rep CanState x

to :: Rep CanState x -> CanState

Generic CanisterContent

Instance details

Associated Types

type Rep CanisterContent :: Type -> Type

Methods

from :: CanisterContent -> Rep CanisterContent x

to :: Rep CanisterContent x -> CanisterContent

Generic CanisterHistory

Instance details

Associated Types

type Rep CanisterHistory :: Type -> Type

Methods

from :: CanisterHistory -> Rep CanisterHistory x

to :: Rep CanisterHistory x -> CanisterHistory

Generic EntryPoint

Instance details

Associated Types

type Rep EntryPoint :: Type -> Type

Methods

from :: EntryPoint -> Rep EntryPoint x

to :: Rep EntryPoint x -> EntryPoint

Generic IC

Instance details

Associated Types

type Rep IC :: Type -> Type

Methods

from :: IC -> Rep IC x

to :: Rep IC x -> IC

Generic Message

Instance details

Associated Types

type Rep Message :: Type -> Type

Methods

from :: Message -> Rep Message x

to :: Rep Message x -> Message

Generic RequestStatus

Instance details

Associated Types

type Rep RequestStatus :: Type -> Type

Methods

from :: RequestStatus -> Rep RequestStatus x

to :: Rep RequestStatus x -> RequestStatus

Generic RunStatus

Instance details

Associated Types

type Rep RunStatus :: Type -> Type

Methods

from :: RunStatus -> Rep RunStatus x

to :: Rep RunStatus x -> RunStatus

Generic Callback

Instance details

Associated Types

type Rep Callback :: Type -> Type

Methods

from :: Callback -> Rep Callback x

to :: Rep Callback x -> Callback

Generic CanisterInstallMode

Instance details

Associated Types

type Rep CanisterInstallMode :: Type -> Type

Methods

from :: CanisterInstallMode -> Rep CanisterInstallMode x

to :: Rep CanisterInstallMode x -> CanisterInstallMode

Generic Change

Instance details

Associated Types

type Rep Change :: Type -> Type

Methods

from :: Change -> Rep Change x

to :: Rep Change x -> Change

Generic ChangeDetails

Instance details

Associated Types

type Rep ChangeDetails :: Type -> Type

Methods

from :: ChangeDetails -> Rep ChangeDetails x

to :: Rep ChangeDetails x -> ChangeDetails

Generic ChangeOrigin

Instance details

Associated Types

type Rep ChangeOrigin :: Type -> Type

Methods

from :: ChangeOrigin -> Rep ChangeOrigin x

to :: Rep ChangeOrigin x -> ChangeOrigin

Generic ErrorCode

Instance details

Associated Types

type Rep ErrorCode :: Type -> Type

Methods

from :: ErrorCode -> Rep ErrorCode x

to :: Rep ErrorCode x -> ErrorCode

Generic MethodCall

Instance details

Associated Types

type Rep MethodCall :: Type -> Type

Methods

from :: MethodCall -> Rep MethodCall x

to :: Rep MethodCall x -> MethodCall

Generic NeedsToRespond

Instance details

Associated Types

type Rep NeedsToRespond :: Type -> Type

Methods

from :: NeedsToRespond -> Rep NeedsToRespond x

to :: Rep NeedsToRespond x -> NeedsToRespond

Generic RejectCode

Instance details

Associated Types

type Rep RejectCode :: Type -> Type

Methods

from :: RejectCode -> Rep RejectCode x

to :: Rep RejectCode x -> RejectCode

Generic Response

Instance details

Associated Types

type Rep Response :: Type -> Type

Methods

from :: Response -> Rep Response x

to :: Rep Response x -> Response

Generic SubnetType

Instance details

Associated Types

type Rep SubnetType :: Type -> Type

Methods

from :: SubnetType -> Rep SubnetType x

to :: Rep SubnetType x -> SubnetType

Generic Timestamp

Instance details

Associated Types

type Rep Timestamp :: Type -> Type

Methods

from :: Timestamp -> Rep Timestamp x

to :: Rep Timestamp x -> Timestamp

Generic WasmClosure

Instance details

Associated Types

type Rep WasmClosure :: Type -> Type

Methods

from :: WasmClosure -> Rep WasmClosure x

to :: Rep WasmClosure x -> WasmClosure

Generic PInstance

Instance details

Associated Types

type Rep PInstance :: Type -> Type

Methods

from :: PInstance -> Rep PInstance x

to :: Rep PInstance x -> PInstance

Generic PModuleInst

Instance details

Associated Types

type Rep PModuleInst :: Type -> Type

Methods

from :: PModuleInst -> Rep PModuleInst x

to :: Rep PModuleInst x -> PModuleInst

ToJSON (Replay i)

Instance details

Methods

toJSON :: Replay i -> Value

toEncoding :: Replay i -> Encoding

toJSONList :: [Replay i] -> Value

toEncodingList :: [Replay i] -> Encoding

ToJSON a => ToJSON (Snapshot a)

Instance details

Methods

toJSON :: Snapshot a -> Value

toEncoding :: Snapshot a -> Encoding

toJSONList :: [Snapshot a] -> Value

toEncodingList :: [Snapshot a] -> Encoding

ToJSON (Module f)

Instance details

Methods

toJSON :: Module f -> Value

toEncoding :: Module f -> Encoding

toJSONList :: [Module f] -> Value

toEncodingList :: [Module f] -> Encoding

Generic (Snapshot a)

Instance details

Associated Types

type Rep (Snapshot a) :: Type -> Type

Methods

from :: Snapshot a -> Rep (Snapshot a) x

to :: Rep (Snapshot a) x -> Snapshot a