IC.Debug.JSON

Description

This module defines ToJSON instances of the IC state.

We put them into their own module, despite the usual advise against orphan instances, to emphasize that these are there just for debugging purposes using `ic-ref`.

(Why JSON? Because Browsers render them nicely in a interactive display where you can open and collapse subcomponents – much easier to get this feature this way, compared to writing custom HTML output.)

Orphan instances

Generic MethodCall

Instance details

Associated Types

type Rep MethodCall :: Type -> Type

Methods

from :: MethodCall -> Rep MethodCall x

to :: Rep MethodCall x -> MethodCall

Generic Callback

Instance details

Associated Types

type Rep Callback :: Type -> Type

Methods

from :: Callback -> Rep Callback x

to :: Rep Callback x -> Callback

Generic WasmClosure

Instance details

Associated Types

type Rep WasmClosure :: Type -> Type

Methods

from :: WasmClosure -> Rep WasmClosure x

to :: Rep WasmClosure x -> WasmClosure

Generic Response

Instance details

Associated Types

type Rep Response :: Type -> Type

Methods

from :: Response -> Rep Response x

to :: Rep Response x -> Response

Generic RejectCode

Instance details

Associated Types

type Rep RejectCode :: Type -> Type

Methods

from :: RejectCode -> Rep RejectCode x

to :: Rep RejectCode x -> RejectCode

Generic Timestamp

Instance details

Associated Types

type Rep Timestamp :: Type -> Type

Methods

from :: Timestamp -> Rep Timestamp x

to :: Rep Timestamp x -> Timestamp

Generic Responded

Instance details

Associated Types

type Rep Responded :: Type -> Type

Methods

from :: Responded -> Rep Responded x

to :: Rep Responded x -> Responded

Generic PModuleInst

Instance details

Associated Types

type Rep PModuleInst :: Type -> Type

Methods

from :: PModuleInst -> Rep PModuleInst x

to :: Rep PModuleInst x -> PModuleInst

Generic PInstance

Instance details

Associated Types

type Rep PInstance :: Type -> Type

Methods

from :: PInstance -> Rep PInstance x

to :: Rep PInstance x -> PInstance

Generic CanisterSnapshot

Instance details

Associated Types

type Rep CanisterSnapshot :: Type -> Type

Methods

from :: CanisterSnapshot -> Rep CanisterSnapshot x

to :: Rep CanisterSnapshot x -> CanisterSnapshot

Generic IC

Instance details

Associated Types

type Rep IC :: Type -> Type

Methods

from :: IC -> Rep IC x

to :: Rep IC x -> IC

Generic Message

Instance details

Associated Types

type Rep Message :: Type -> Type

Methods

from :: Message -> Rep Message x

to :: Rep Message x -> Message

Generic CallOrigin

Instance details

Associated Types

type Rep CallOrigin :: Type -> Type

Methods

from :: CallOrigin -> Rep CallOrigin x

to :: Rep CallOrigin x -> CallOrigin

Generic CallContext

Instance details

Associated Types

type Rep CallContext :: Type -> Type

Methods

from :: CallContext -> Rep CallContext x

to :: Rep CallContext x -> CallContext

Generic EntryPoint

Instance details

Associated Types

type Rep EntryPoint :: Type -> Type

Methods

from :: EntryPoint -> Rep EntryPoint x

to :: Rep EntryPoint x -> EntryPoint

Generic CanState

Instance details

Associated Types

type Rep CanState :: Type -> Type

Methods

from :: CanState -> Rep CanState x

to :: Rep CanState x -> CanState

Generic CanisterContent

Instance details

Associated Types

type Rep CanisterContent :: Type -> Type

Methods

from :: CanisterContent -> Rep CanisterContent x

to :: Rep CanisterContent x -> CanisterContent

Generic RunStatus

Instance details

Associated Types

type Rep RunStatus :: Type -> Type

Methods

from :: RunStatus -> Rep RunStatus x

to :: Rep RunStatus x -> RunStatus

Generic CallResponse

Instance details

Associated Types

type Rep CallResponse :: Type -> Type

Methods

from :: CallResponse -> Rep CallResponse x

to :: Rep CallResponse x -> CallResponse

Generic RequestStatus

Instance details

Associated Types

type Rep RequestStatus :: Type -> Type

Methods

from :: RequestStatus -> Rep RequestStatus x

to :: Rep RequestStatus x -> RequestStatus

Generic CallRequest

Instance details

Associated Types

type Rep CallRequest :: Type -> Type

Methods

from :: CallRequest -> Rep CallRequest x

to :: Rep CallRequest x -> CallRequest

ToJSON ByteString

Instance details

Methods

toJSON :: ByteString -> Value

toEncoding :: ByteString -> Encoding

toJSONList :: [ByteString] -> Value

toEncodingList :: [ByteString] -> Encoding

ToJSON StdGen

Instance details

Methods

toJSON :: StdGen -> Value

toEncoding :: StdGen -> Encoding

toJSONList :: [StdGen] -> Value

toEncodingList :: [StdGen] -> Encoding

ToJSON MethodCall

Instance details

Methods

toJSON :: MethodCall -> Value

toEncoding :: MethodCall -> Encoding

toJSONList :: [MethodCall] -> Value

toEncodingList :: [MethodCall] -> Encoding

ToJSON Callback

Instance details

Methods

toJSON :: Callback -> Value

toEncoding :: Callback -> Encoding

toJSONList :: [Callback] -> Value

toEncodingList :: [Callback] -> Encoding

ToJSON WasmClosure

Instance details

Methods

toJSON :: WasmClosure -> Value

toEncoding :: WasmClosure -> Encoding

toJSONList :: [WasmClosure] -> Value

toEncodingList :: [WasmClosure] -> Encoding

ToJSON Response

Instance details

Methods

toJSON :: Response -> Value

toEncoding :: Response -> Encoding

toJSONList :: [Response] -> Value

toEncodingList :: [Response] -> Encoding

ToJSON RejectCode

Instance details

Methods

toJSON :: RejectCode -> Value

toEncoding :: RejectCode -> Encoding

toJSONList :: [RejectCode] -> Value

toEncodingList :: [RejectCode] -> Encoding

ToJSON Timestamp

Instance details

Methods

toJSON :: Timestamp -> Value

toEncoding :: Timestamp -> Encoding

toJSONList :: [Timestamp] -> Value

toEncodingList :: [Timestamp] -> Encoding

ToJSON Responded

Instance details

Methods

toJSON :: Responded -> Value

toEncoding :: Responded -> Encoding

toJSONList :: [Responded] -> Value

toEncodingList :: [Responded] -> Encoding

ToJSON EntityId

Instance details

Methods

toJSON :: EntityId -> Value

toEncoding :: EntityId -> Encoding

toJSONList :: [EntityId] -> Value

toEncodingList :: [EntityId] -> Encoding

ToJSON SecretKey

Instance details

Methods

toJSON :: SecretKey -> Value

toEncoding :: SecretKey -> Encoding

toJSONList :: [SecretKey] -> Value

toEncodingList :: [SecretKey] -> Encoding

ToJSON Value

Instance details

Methods

toJSON :: Value -> Value0

toEncoding :: Value -> Encoding

toJSONList :: [Value] -> Value0

toEncodingList :: [Value] -> Encoding

ToJSON PModuleInst

Instance details

Methods

toJSON :: PModuleInst -> Value

toEncoding :: PModuleInst -> Encoding

toJSONList :: [PModuleInst] -> Value

toEncodingList :: [PModuleInst] -> Encoding

ToJSON PInstance

Instance details

Methods

toJSON :: PInstance -> Value

toEncoding :: PInstance -> Encoding

toJSONList :: [PInstance] -> Value

toEncodingList :: [PInstance] -> Encoding

ToJSON CanisterSnapshot

Instance details

Methods

toJSON :: CanisterSnapshot -> Value

toEncoding :: CanisterSnapshot -> Encoding

toJSONList :: [CanisterSnapshot] -> Value

toEncodingList :: [CanisterSnapshot] -> Encoding

ToJSON CanisterModule

Instance details

Methods

toJSON :: CanisterModule -> Value

toEncoding :: CanisterModule -> Encoding

toJSONList :: [CanisterModule] -> Value

toEncodingList :: [CanisterModule] -> Encoding

ToJSON IC

Instance details

Methods

toJSON :: IC -> Value

toEncoding :: IC -> Encoding

toJSONList :: [IC] -> Value

toEncodingList :: [IC] -> Encoding

ToJSON Message

Instance details

Methods

toJSON :: Message -> Value

toEncoding :: Message -> Encoding

toJSONList :: [Message] -> Value

toEncodingList :: [Message] -> Encoding

ToJSON CallOrigin

Instance details

Methods

toJSON :: CallOrigin -> Value

toEncoding :: CallOrigin -> Encoding

toJSONList :: [CallOrigin] -> Value

toEncodingList :: [CallOrigin] -> Encoding

ToJSON CallContext

Instance details

Methods

toJSON :: CallContext -> Value

toEncoding :: CallContext -> Encoding

toJSONList :: [CallContext] -> Value

toEncodingList :: [CallContext] -> Encoding

ToJSON EntryPoint

Instance details

Methods

toJSON :: EntryPoint -> Value

toEncoding :: EntryPoint -> Encoding

toJSONList :: [EntryPoint] -> Value

toEncodingList :: [EntryPoint] -> Encoding

ToJSON CanState

Instance details

Methods

toJSON :: CanState -> Value

toEncoding :: CanState -> Encoding

toJSONList :: [CanState] -> Value

toEncodingList :: [CanState] -> Encoding

ToJSON CanisterContent

Instance details

Methods

toJSON :: CanisterContent -> Value

toEncoding :: CanisterContent -> Encoding

toJSONList :: [CanisterContent] -> Value

toEncodingList :: [CanisterContent] -> Encoding

ToJSON RunStatus

Instance details

Methods

toJSON :: RunStatus -> Value

toEncoding :: RunStatus -> Encoding

toJSONList :: [RunStatus] -> Value

toEncodingList :: [RunStatus] -> Encoding

ToJSON CallResponse

Instance details

Methods

toJSON :: CallResponse -> Value

toEncoding :: CallResponse -> Encoding

toJSONList :: [CallResponse] -> Value

toEncodingList :: [CallResponse] -> Encoding

ToJSON RequestStatus

Instance details

Methods

toJSON :: RequestStatus -> Value

toEncoding :: RequestStatus -> Encoding

toJSONList :: [RequestStatus] -> Value

toEncodingList :: [RequestStatus] -> Encoding

ToJSON CallRequest

Instance details

Methods

toJSON :: CallRequest -> Value

toEncoding :: CallRequest -> Encoding

toJSONList :: [CallRequest] -> Value

toEncodingList :: [CallRequest] -> Encoding

ToJSONKey ByteString

Instance details

Methods

toJSONKey :: ToJSONKeyFunction ByteString

toJSONKeyList :: ToJSONKeyFunction [ByteString]

ToJSONKey EntityId

Instance details

Methods

toJSONKey :: ToJSONKeyFunction EntityId

toJSONKeyList :: ToJSONKeyFunction [EntityId]

Generic (Snapshot a)

Instance details

Associated Types

type Rep (Snapshot a) :: Type -> Type

Methods

from :: Snapshot a -> Rep (Snapshot a) x

to :: Rep (Snapshot a) x -> Snapshot a

ToJSON (Replay i)

Instance details

Methods

toJSON :: Replay i -> Value

toEncoding :: Replay i -> Encoding

toJSONList :: [Replay i] -> Value

toEncodingList :: [Replay i] -> Encoding

ToJSON a => ToJSON (Snapshot a)

Instance details

Methods

toJSON :: Snapshot a -> Value

toEncoding :: Snapshot a -> Encoding

toJSONList :: [Snapshot a] -> Value

toEncodingList :: [Snapshot a] -> Encoding

ToJSON (Module f)

Instance details

Methods

toJSON :: Module f -> Value

toEncoding :: Module f -> Encoding

toJSONList :: [Module f] -> Value

toEncodingList :: [Module f] -> Encoding