IC.Ref

Description

This module implements the main abstract logic of the Internet Computer. It assumes a pure and abstracted view on Canisters (provided by IC.Canister), and deals with abstract requests (CallRequest, QueryRequest, ...), so HTTP and CBOR-level processing has already happened.

Documentation

data IC

Constructors

IC

Fields

canisters :: CanisterId ↦ CanState

requests :: RequestID ↦ (CallRequest, RequestStatus)

messages :: Seq Message

call_contexts :: CallId ↦ CallContext

rng :: StdGen

secretRootKey :: SecretKey

secretSubnetKey :: SecretKey

data CallRequest

Constructors

CallRequest CanisterId UserId MethodName Blob

callerOfCallRequest :: CallRequest -> EntityId

data QueryRequest

Constructors

QueryRequest CanisterId UserId MethodName Blob

data ReadStateRequest

Constructors

ReadStateRequest UserId [Path]

data RequestStatus

Constructors

Received

Processing

CallResponse CallResponse

data ReqResponse

Constructors

QueryResponse CallResponse

ReadStateResponse Certificate

data CallResponse

Constructors

Rejected (RejectCode, String)

Replied Blob

initialIC :: IO IC

authCallRequest :: RequestValidation m => Timestamp -> CanisterId -> EnvValidity -> CallRequest -> m ()

authQueryRequest :: RequestValidation m => Timestamp -> CanisterId -> EnvValidity -> QueryRequest -> m ()

authReadStateRequest :: RequestValidation m => Timestamp -> CanisterId -> EnvValidity -> ReadStateRequest -> m ()

submitRequest :: ICM m => RequestID -> CallRequest -> m ()

Submission simply enqueues requests

handleQuery :: ICM m => Timestamp -> QueryRequest -> m ReqResponse

handleReadState :: ICM m => Timestamp -> ReadStateRequest -> m ReqResponse

runStep :: ICM m => m Bool

Returns true if a step was taken

runToCompletion :: ICM m => m ()

setAllTimesTo :: ICM m => Timestamp -> m ()

createEmptyCanister :: ICM m => CanisterId -> Set EntityId -> Timestamp -> m ()

data CallContext

Constructors

CallContext

Fields

canister :: CanisterId

origin :: CallOrigin

responded :: Responded

deleted :: Bool

available_cycles :: Cycles

last_trap :: Maybe String

non-normative, but yields better reject messages

data Message

Constructors

CallMessage

Fields

call_context :: CallId

entry :: EntryPoint

ResponseMessage

Fields

call_context :: CallId

response :: Response

refunded_cycles :: Cycles

data CanState

Constructors

CanState

Fields

content :: Maybe CanisterContent

run_status :: RunStatus

controllers :: Set EntityId

memory_allocation :: Natural

compute_allocation :: Natural

freezing_threshold :: Natural

time :: Timestamp

cycle_balance :: Natural

certified_data :: Blob

data CallOrigin

Constructors

FromUser RequestID

FromCanister CallId Callback

data EntryPoint

Constructors

Public MethodName Blob

Closure Callback Response Cycles

data RunStatus

Constructors

IsRunning

IsStopping [CallId]

IsStopped

IsDeleted

data CanisterContent

Constructors

CanisterContent

Fields

can_mod :: CanisterModule

wasm_state :: WasmState