IC.Serialise

Description

This module defines Serialise instances of the IC state.

We put them into their own module, despite the usual advise against orphan instances, to emphasize that these are not part of the Ref module with its “reference implementation” status.

Also, orphan instances are kinda ok in applications.

Orphan instances

Generic CallContext

Instance details

Associated Types

type Rep CallContext :: Type -> Type

Methods

from :: CallContext -> Rep CallContext x

to :: Rep CallContext x -> CallContext

Generic CallOrigin

Instance details

Associated Types

type Rep CallOrigin :: Type -> Type

Methods

from :: CallOrigin -> Rep CallOrigin x

to :: Rep CallOrigin x -> CallOrigin

Generic CallRequest

Instance details

Associated Types

type Rep CallRequest :: Type -> Type

Methods

from :: CallRequest -> Rep CallRequest x

to :: Rep CallRequest x -> CallRequest

Generic CallResponse

Instance details

Associated Types

type Rep CallResponse :: Type -> Type

Methods

from :: CallResponse -> Rep CallResponse x

to :: Rep CallResponse x -> CallResponse

Generic CanState

Instance details

Associated Types

type Rep CanState :: Type -> Type

Methods

from :: CanState -> Rep CanState x

to :: Rep CanState x -> CanState

Generic CanisterHistory

Instance details

Associated Types

type Rep CanisterHistory :: Type -> Type

Methods

from :: CanisterHistory -> Rep CanisterHistory x

to :: Rep CanisterHistory x -> CanisterHistory

Generic EntryPoint

Instance details

Associated Types

type Rep EntryPoint :: Type -> Type

Methods

from :: EntryPoint -> Rep EntryPoint x

to :: Rep EntryPoint x -> EntryPoint

Generic IC

Instance details

Associated Types

type Rep IC :: Type -> Type

Methods

from :: IC -> Rep IC x

to :: Rep IC x -> IC

Generic Message

Instance details

Associated Types

type Rep Message :: Type -> Type

Methods

from :: Message -> Rep Message x

to :: Rep Message x -> Message

Generic RequestStatus

Instance details

Associated Types

type Rep RequestStatus :: Type -> Type

Methods

from :: RequestStatus -> Rep RequestStatus x

to :: Rep RequestStatus x -> RequestStatus

Generic RunStatus

Instance details

Associated Types

type Rep RunStatus :: Type -> Type

Methods

from :: RunStatus -> Rep RunStatus x

to :: Rep RunStatus x -> RunStatus

Generic Callback

Instance details

Associated Types

type Rep Callback :: Type -> Type

Methods

from :: Callback -> Rep Callback x

to :: Rep Callback x -> Callback

Generic CanisterInstallMode

Instance details

Associated Types

type Rep CanisterInstallMode :: Type -> Type

Methods

from :: CanisterInstallMode -> Rep CanisterInstallMode x

to :: Rep CanisterInstallMode x -> CanisterInstallMode

Generic Change

Instance details

Associated Types

type Rep Change :: Type -> Type

Methods

from :: Change -> Rep Change x

to :: Rep Change x -> Change

Generic ChangeDetails

Instance details

Associated Types

type Rep ChangeDetails :: Type -> Type

Methods

from :: ChangeDetails -> Rep ChangeDetails x

to :: Rep ChangeDetails x -> ChangeDetails

Generic ChangeOrigin

Instance details

Associated Types

type Rep ChangeOrigin :: Type -> Type

Methods

from :: ChangeOrigin -> Rep ChangeOrigin x

to :: Rep ChangeOrigin x -> ChangeOrigin

Generic ErrorCode

Instance details

Associated Types

type Rep ErrorCode :: Type -> Type

Methods

from :: ErrorCode -> Rep ErrorCode x

to :: Rep ErrorCode x -> ErrorCode

Generic MethodCall

Instance details

Associated Types

type Rep MethodCall :: Type -> Type

Methods

from :: MethodCall -> Rep MethodCall x

to :: Rep MethodCall x -> MethodCall

Generic NeedsToRespond

Instance details

Associated Types

type Rep NeedsToRespond :: Type -> Type

Methods

from :: NeedsToRespond -> Rep NeedsToRespond x

to :: Rep NeedsToRespond x -> NeedsToRespond

Generic RejectCode

Instance details

Associated Types

type Rep RejectCode :: Type -> Type

Methods

from :: RejectCode -> Rep RejectCode x

to :: Rep RejectCode x -> RejectCode

Generic Response

Instance details

Associated Types

type Rep Response :: Type -> Type

Methods

from :: Response -> Rep Response x

to :: Rep Response x -> Response

Generic SubnetType

Instance details

Associated Types

type Rep SubnetType :: Type -> Type

Methods

from :: SubnetType -> Rep SubnetType x

to :: Rep SubnetType x -> SubnetType

Generic Timestamp

Instance details

Associated Types

type Rep Timestamp :: Type -> Type

Methods

from :: Timestamp -> Rep Timestamp x

to :: Rep Timestamp x -> Timestamp

Generic WasmClosure

Instance details

Associated Types

type Rep WasmClosure :: Type -> Type

Methods

from :: WasmClosure -> Rep WasmClosure x

to :: Rep WasmClosure x -> WasmClosure

Generic PInstance

Instance details

Associated Types

type Rep PInstance :: Type -> Type

Methods

from :: PInstance -> Rep PInstance x

to :: Rep PInstance x -> PInstance

Generic PModuleInst

Instance details

Associated Types

type Rep PModuleInst :: Type -> Type

Methods

from :: PModuleInst -> Rep PModuleInst x

to :: Rep PModuleInst x -> PModuleInst

Serialise Repr

Instance details

Methods

encode :: Repr -> Encoding

decode :: Decoder s Repr

encodeList :: [Repr] -> Encoding

decodeList :: Decoder s [Repr]

Serialise SecretKey

Instance details

Methods

encode :: SecretKey -> Encoding

decode :: Decoder s SecretKey

encodeList :: [SecretKey] -> Encoding

decodeList :: Decoder s [SecretKey]

Serialise SecretKey

Instance details

Methods

encode :: SecretKey -> Encoding

decode :: Decoder s SecretKey

encodeList :: [SecretKey] -> Encoding

decodeList :: Decoder s [SecretKey]

Serialise CallContext

Instance details

Methods

encode :: CallContext -> Encoding

decode :: Decoder s CallContext

encodeList :: [CallContext] -> Encoding

decodeList :: Decoder s [CallContext]

Serialise CallOrigin

Instance details

Methods

encode :: CallOrigin -> Encoding

decode :: Decoder s CallOrigin

encodeList :: [CallOrigin] -> Encoding

decodeList :: Decoder s [CallOrigin]

Serialise CallRequest

Instance details

Methods

encode :: CallRequest -> Encoding

decode :: Decoder s CallRequest

encodeList :: [CallRequest] -> Encoding

decodeList :: Decoder s [CallRequest]

Serialise CallResponse

Instance details

Methods

encode :: CallResponse -> Encoding

decode :: Decoder s CallResponse

encodeList :: [CallResponse] -> Encoding

decodeList :: Decoder s [CallResponse]

Serialise CanState

Instance details

Methods

encode :: CanState -> Encoding

decode :: Decoder s CanState

encodeList :: [CanState] -> Encoding

decodeList :: Decoder s [CanState]

Serialise CanisterContent

Instance details

Methods

encode :: CanisterContent -> Encoding

decode :: Decoder s CanisterContent

encodeList :: [CanisterContent] -> Encoding

decodeList :: Decoder s [CanisterContent]

Serialise CanisterHistory

Instance details

Methods

encode :: CanisterHistory -> Encoding

decode :: Decoder s CanisterHistory

encodeList :: [CanisterHistory] -> Encoding

decodeList :: Decoder s [CanisterHistory]

Serialise EntryPoint

Instance details

Methods

encode :: EntryPoint -> Encoding

decode :: Decoder s EntryPoint

encodeList :: [EntryPoint] -> Encoding

decodeList :: Decoder s [EntryPoint]

Serialise IC

Instance details

Methods

encode :: IC -> Encoding

decode :: Decoder s IC

encodeList :: [IC] -> Encoding

decodeList :: Decoder s [IC]

Serialise Message

Instance details

Methods

encode :: Message -> Encoding

decode :: Decoder s Message

encodeList :: [Message] -> Encoding

decodeList :: Decoder s [Message]

Serialise RequestStatus

Instance details

Methods

encode :: RequestStatus -> Encoding

decode :: Decoder s RequestStatus

encodeList :: [RequestStatus] -> Encoding

decodeList :: Decoder s [RequestStatus]

Serialise RunStatus

Instance details

Methods

encode :: RunStatus -> Encoding

decode :: Decoder s RunStatus

encodeList :: [RunStatus] -> Encoding

decodeList :: Decoder s [RunStatus]

Serialise Callback

Instance details

Methods

encode :: Callback -> Encoding

decode :: Decoder s Callback

encodeList :: [Callback] -> Encoding

decodeList :: Decoder s [Callback]

Serialise CanisterInstallMode

Instance details

Methods

encode :: CanisterInstallMode -> Encoding

decode :: Decoder s CanisterInstallMode

encodeList :: [CanisterInstallMode] -> Encoding

decodeList :: Decoder s [CanisterInstallMode]

Serialise Change

Instance details

Methods

encode :: Change -> Encoding

decode :: Decoder s Change

encodeList :: [Change] -> Encoding

decodeList :: Decoder s [Change]

Serialise ChangeDetails

Instance details

Methods

encode :: ChangeDetails -> Encoding

decode :: Decoder s ChangeDetails

encodeList :: [ChangeDetails] -> Encoding

decodeList :: Decoder s [ChangeDetails]

Serialise ChangeOrigin

Instance details

Methods

encode :: ChangeOrigin -> Encoding

decode :: Decoder s ChangeOrigin

encodeList :: [ChangeOrigin] -> Encoding

decodeList :: Decoder s [ChangeOrigin]

Serialise EntityId

Instance details

Methods

encode :: EntityId -> Encoding

decode :: Decoder s EntityId

encodeList :: [EntityId] -> Encoding

decodeList :: Decoder s [EntityId]

Serialise ErrorCode

Instance details

Methods

encode :: ErrorCode -> Encoding

decode :: Decoder s ErrorCode

encodeList :: [ErrorCode] -> Encoding

decodeList :: Decoder s [ErrorCode]

Serialise MethodCall

Instance details

Methods

encode :: MethodCall -> Encoding

decode :: Decoder s MethodCall

encodeList :: [MethodCall] -> Encoding

decodeList :: Decoder s [MethodCall]

Serialise NeedsToRespond

Instance details

Methods

encode :: NeedsToRespond -> Encoding

decode :: Decoder s NeedsToRespond

encodeList :: [NeedsToRespond] -> Encoding

decodeList :: Decoder s [NeedsToRespond]

Serialise RejectCode

Instance details

Methods

encode :: RejectCode -> Encoding

decode :: Decoder s RejectCode

encodeList :: [RejectCode] -> Encoding

decodeList :: Decoder s [RejectCode]

Serialise Response

Instance details

Methods

encode :: Response -> Encoding

decode :: Decoder s Response

encodeList :: [Response] -> Encoding

decodeList :: Decoder s [Response]

Serialise SubnetType

Instance details

Methods

encode :: SubnetType -> Encoding

decode :: Decoder s SubnetType

encodeList :: [SubnetType] -> Encoding

decodeList :: Decoder s [SubnetType]

Serialise Timestamp

Instance details

Methods

encode :: Timestamp -> Encoding

decode :: Decoder s Timestamp

encodeList :: [Timestamp] -> Encoding

decodeList :: Decoder s [Timestamp]

Serialise WasmClosure

Instance details

Methods

encode :: WasmClosure -> Encoding

decode :: Decoder s WasmClosure

encodeList :: [WasmClosure] -> Encoding

decodeList :: Decoder s [WasmClosure]

Serialise PInstance

Instance details

Methods

encode :: PInstance -> Encoding

decode :: Decoder s PInstance

encodeList :: [PInstance] -> Encoding

decodeList :: Decoder s [PInstance]

Serialise PModuleInst

Instance details

Methods

encode :: PModuleInst -> Encoding

decode :: Decoder s PModuleInst

encodeList :: [PModuleInst] -> Encoding

decodeList :: Decoder s [PModuleInst]

Serialise StdGen

Instance details

Methods

encode :: StdGen -> Encoding

decode :: Decoder s StdGen

encodeList :: [StdGen] -> Encoding

decodeList :: Decoder s [StdGen]

Serialise SMGen

Instance details

Methods

encode :: SMGen -> Encoding

decode :: Decoder s SMGen

encodeList :: [SMGen] -> Encoding

decodeList :: Decoder s [SMGen]

Serialise Value

Instance details

Methods

encode :: Value -> Encoding

decode :: Decoder s Value

encodeList :: [Value] -> Encoding

decodeList :: Decoder s [Value]

Generic (Snapshot a)

Instance details

Associated Types

type Rep (Snapshot a) :: Type -> Type

Methods

from :: Snapshot a -> Rep (Snapshot a) x

to :: Rep (Snapshot a) x -> Snapshot a

Serialise a => Serialise (Snapshot a)

Instance details

Methods

encode :: Snapshot a -> Encoding

decode :: Decoder s (Snapshot a)

encodeList :: [Snapshot a] -> Encoding

decodeList :: Decoder s [Snapshot a]