IC.Test.Agent

Description

This module contains function to interact with an Internet Computer instance.

The primary customer here is IC.Test.Spec, i.e the test suite. Therefore, the functions here give access to varoius levels of abstractions, and gives more control than a “normal” agent library would do.

Also, because of the focus on testing, failures are repoted directly with HUnit’s assertFailure.

As guidance: This modules does _not_ rely on the universal canister.

Documentation

type HTTPErrOr a = Either (Int, String) a

type HasAgentConfig = ?agentConfig :: AgentConfig

type IC00 = Blob -> Text -> Blob -> IO ReqResponse

type IC00WithCycles = Word64 -> IC00

type IC00' = Blob -> Text -> Blob -> IO (HTTPErrOr ReqResponse)

data ReqResponse

Constructors

Reply Blob

Reject Natural Text (Maybe Text)

data ReqStatus

Constructors

Processing

Pending

Responded ReqResponse

UnknownStatus

data AgentConfig

Constructors

AgentConfig

Fields

tc_root_key :: Blob

tc_manager :: Manager

tc_endPoint :: String

tc_subnets :: [AgentSubnetConfig]

tc_httpbin :: String

tc_timeout :: Int

data DelegationCanisterRangeCheck

Constructors

DelegationCanisterRangeCheck [(Blob, Blob)] Blob

addExpiry :: GenR -> IO GenR

addNonce :: GenR -> IO GenR

addNonceExpiryEnv :: GenR -> IO GenR

anonymousUser :: Blob

as2Word64 :: HasCallStack => Blob -> IO (Word64, Word64)

asWord64Word128 :: HasCallStack => Blob -> IO (Word64, Word128)

asHex :: Blob -> String

asRight :: HasCallStack => Either Text a -> IO a

asWord128 :: HasCallStack => Blob -> IO Word128

asWord32 :: HasCallStack => Blob -> IO Word32

asWord64 :: HasCallStack => Blob -> IO Word64

awaitCall' :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (HTTPErrOr ReqResponse)

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit". Returns either a HTTP Error code, or if the status is 2xx, poll for the request response, and return decoded CBOR

awaitCall :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO ReqResponse

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit", poll for the request response, and return decoded CBOR

awaitKnown :: HasAgentConfig => IO (HTTPErrOr ReqStatus) -> IO ReqStatus

awaitStatus :: HasAgentConfig => IO (HTTPErrOr ReqStatus) -> IO ReqResponse

bothSame :: (Eq a, Show a) => (a, a) -> Assertion

callResponse :: GenR -> IO ReqResponse

certValue :: HasCallStack => CertVal a => Certificate -> [Blob] -> IO a

certValueAbsent :: HasCallStack => Certificate -> [Blob] -> IO ()

code202 :: HasCallStack => Response ByteString -> IO ()

code202_or_4xx :: HasCallStack => Response ByteString -> IO ()

code2xx :: HasCallStack => Response ByteString -> IO ()

code4xx :: HasCallStack => Response ByteString -> IO ()

decodeCert' :: HasCallStack => Blob -> IO Certificate

defaultSK :: SecretKey

defaultUser :: Blob

delegationEnv :: SecretKey -> [(SecretKey, Maybe [Blob])] -> GenR -> IO GenR

doesn'tExist :: Blob

ecdsaSK :: SecretKey

ecdsaUser :: Blob

enum :: (AllUniqueLabels r, KnownSymbol l, (r .! l) ~ ()) => Label l -> Var r

envelope :: SecretKey -> GenR -> IO GenR

envelopeFor :: Blob -> GenR -> IO GenR

extractCertData :: Blob -> Blob -> IO Blob

getRequestStatus' :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Blob -> IO (HTTPErrOr ReqStatus)

getRequestStatus :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Blob -> IO ReqStatus

getStateCert' :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO (HTTPErrOr Certificate)

getStateCert :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO Certificate

ic00 :: HasAgentConfig => IC00

ic00as :: (HasAgentConfig, HasCallStack) => Blob -> IC00

ic00' :: HasAgentConfig => IC00'

ic00WithSubnetas' :: HasAgentConfig => Blob -> Blob -> Blob -> Text -> Blob -> IO (HTTPErrOr ReqResponse)

ingressDelay :: IO ()

is2xx :: HasCallStack => HTTPErrOr a -> IO a

isErr4xx :: HasCallStack => HTTPErrOr a -> IO ()

isErrOrReject :: HasCallStack => [Natural] -> HTTPErrOr ReqResponse -> IO ()

isNoErrReject :: HasCallStack => [Natural] -> HTTPErrOr ReqResponse -> IO ()

isPendingOrProcessing :: ReqStatus -> IO ()

isReject :: HasCallStack => [Natural] -> ReqResponse -> IO ()

isReply :: HasCallStack => ReqResponse -> IO Blob

isResponded :: ReqStatus -> Assertion

okCBOR :: HasCallStack => Response ByteString -> IO GenR

otherSK :: SecretKey

otherUser :: Blob

makeAgentConfig :: Bool -> String -> [AgentSubnetConfig] -> String -> Int -> IO AgentConfig

postCBOR :: (HasCallStack, HasAgentConfig) => String -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

postCallCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postQueryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postReadStateCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

preFlight :: OptionSet -> IO AgentConfig

queryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO GenR

Add envelope to CBOR request, add a nonce and expiry if it is not there, post to "read", return decoded CBOR

queryResponse :: GenR -> IO ReqResponse

runGet :: HasCallStack => Get a -> Blob -> IO a

secp256k1SK :: SecretKey

secp256k1User :: Blob

senderOf :: GenR -> Blob

shorten :: Int -> String -> String

submitCall :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (IO (HTTPErrOr ReqStatus))

textual :: Blob -> String

validateStateCert :: (HasCallStack, HasAgentConfig) => Blob -> Certificate -> IO ()

verifySignature :: Blob -> Blob -> Blob -> Bool

waitFor :: HasAgentConfig => IO (Maybe a) -> IO a

webAuthnECDSASK :: SecretKey

webAuthnECDSAUser :: Blob

webAuthnRSASK :: SecretKey

webAuthnRSAUser :: Blob

withAgentConfig :: (forall. HasAgentConfig => a) -> AgentConfig -> a

callIC :: forall s a b. (HasCallStack, HasAgentConfig) => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => (CandidArg a, CandidArg b) => IC00 -> Blob -> Label s -> a -> IO b

callIC' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => IC00 -> Blob -> Label s -> a -> IO ReqResponse

callIC'' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => Blob -> Blob -> Label s -> a -> IO (HTTPErrOr ReqResponse)

callICWithSubnet'' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => Blob -> Blob -> Blob -> Label s -> a -> IO (HTTPErrOr ReqResponse)

callIC''' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => IC00' -> Blob -> Label s -> a -> IO (HTTPErrOr ReqResponse)

agentConfig :: HasAgentConfig => AgentConfig