IC.Test.Agent

Description

This module contains function to interact with an Internet Computer instance.

The primary customer here is IC.Test.Spec, i.e the test suite. Therefore, the functions here give access to varoius levels of abstractions, and gives more control than a “normal” agent library would do.

Also, because of the focus on testing, failures are repoted directly with HUnit’s assertFailure.

As guidance: This modules does _not_ rely on the universal canister.

This module can also be used in a REPL; see connect.

Agent configuration

data AgentConfig

Constructors

AgentConfig

Fields

tc_root_key :: Blob

tc_manager :: Manager

tc_endPoint :: String

makeAgentConfig :: String -> IO AgentConfig

preFlight :: OptionSet -> IO AgentConfig

newtype ReplWrapper

Constructors

R (forall a. (HasAgentConfig => a) -> a)

connect :: String -> IO ReplWrapper

This is for use from the Haskell REPL, see README.md

type HasAgentConfig = ?agentConfig :: AgentConfig

withAgentConfig :: (forall. HasAgentConfig => a) -> AgentConfig -> a

agentConfig :: HasAgentConfig => AgentConfig

endPoint :: HasAgentConfig => String

agentManager :: HasAgentConfig => Manager

Test data for some hardcoded user names

doesn'tExist :: Blob

defaultSK :: SecretKey

otherSK :: SecretKey

webAuthnECDSASK :: SecretKey

webAuthnRSASK :: SecretKey

ecdsaSK :: SecretKey

secp256k1SK :: SecretKey

defaultUser :: Blob

otherUser :: Blob

webAuthnECDSAUser :: Blob

webAuthnRSAUser :: Blob

ecdsaUser :: Blob

secp256k1User :: Blob

anonymousUser :: Blob

Request envelopes

addIfNotThere :: Monad m => Text -> m GenR -> GenR -> m GenR

addNonce :: GenR -> IO GenR

getRand8Bytes :: IO ByteString

addExpiry :: GenR -> IO GenR

envelope :: SecretKey -> GenR -> IO GenR

delegationEnv :: SecretKey -> [(SecretKey, Maybe [Blob])] -> GenR -> IO GenR

senderOf :: GenR -> Blob

addNonceExpiryEnv :: GenR -> IO GenR

envelopeFor :: Blob -> GenR -> IO GenR

HUnit error reporting integration

asRight :: HasCallStack => Either Text a -> IO a

asExceptT :: HasCallStack => ExceptT Text IO a -> IO a

Requests

postCBOR :: (HasCallStack, HasAgentConfig) => String -> GenR -> IO (Response ByteString)

Posting a CBOR request, returning a raw bytestring

postCallCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

postQueryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

postReadStateCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO (Response ByteString)

postCBOR with url based on effective canister id

queryCBOR :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO GenR

Add envelope to CBOR request, add a nonce and expiry if it is not there, post to "read", return decoded CBOR

type HTTPErrOrReqResponse = Either (Int, String) ReqResponse

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit". Returns either a HTTP Error code, or if the status is 2xx, poll for the request response, and return decoded CBOR

awaitCall' :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO HTTPErrOrReqResponse

awaitCall :: (HasCallStack, HasAgentConfig) => Blob -> GenR -> IO ReqResponse

Add envelope to CBOR, and a nonce and expiry if not there, post to "submit", poll for the request response, and return decoded CBOR

is2xx :: HasCallStack => HTTPErrOrReqResponse -> IO ReqResponse

getStateCert' :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO (Response Blob)

decodeCert' :: HasCallStack => Blob -> IO Certificate

getStateCert :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> [[Blob]] -> IO Certificate

extractCertData :: Blob -> Blob -> IO Blob

verboseVerify :: String -> Blob -> Blob -> Blob -> Blob -> IO ()

validateDelegation :: (HasCallStack, HasAgentConfig) => Maybe Delegation -> IO Blob

validateStateCert' :: (HasCallStack, HasAgentConfig) => String -> Certificate -> IO ()

validateStateCert :: (HasCallStack, HasAgentConfig) => Certificate -> IO ()

data ReqResponse

Constructors

Reply Blob

Reject Natural Text

data ReqStatus

Constructors

Processing

Pending

Responded ReqResponse

UnknownStatus

prettyPath :: [Blob] -> String

prettyBlob :: Blob -> String

certValue :: HasCallStack => CertVal a => Certificate -> [Blob] -> IO a

certValueAbsent :: HasCallStack => Certificate -> [Blob] -> IO ()

getRequestStatus :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Blob -> IO ReqStatus

awaitStatus :: HasAgentConfig => Blob -> Blob -> Blob -> IO ReqResponse

pollDelay :: IO ()

ingressDelay :: IO ()

HTTP Response predicates

codePred :: HasCallStack => String -> (Int -> Bool) -> Response Blob -> IO ()

code2xx :: HasCallStack => Response ByteString -> IO ()

code202 :: HasCallStack => Response ByteString -> IO ()

code4xx :: HasCallStack => Response ByteString -> IO ()

code202_or_4xx :: HasCallStack => Response ByteString -> IO ()

CBOR decoding

okCBOR :: HasCallStack => Response ByteString -> IO GenR

Response predicates and parsers

queryResponse :: GenR -> IO ReqResponse

isReject :: HasCallStack => [Natural] -> ReqResponse -> IO ()

isErrOrReject :: HasCallStack => [Natural] -> HTTPErrOrReqResponse -> IO ()

isReply :: HasCallStack => ReqResponse -> IO Blob

asWord32 :: HasCallStack => Blob -> IO Word32

asWord64 :: HasCallStack => Blob -> IO Word64

as2Word64 :: HasCallStack => Blob -> IO (Word64, Word64)

bothSame :: (Eq a, Show a) => (a, a) -> Assertion

runGet :: HasCallStack => Get a -> Blob -> IO a

Status endpoint parsing

data StatusResponse

Constructors

StatusResponse

Fields

status_api_version :: Text

status_root_key :: Blob

statusResonse :: HasCallStack => GenR -> IO StatusResponse

Interacting with aaaaa-aa (via HTTP)

type IC00 = Blob -> Text -> Blob -> IO ReqResponse

ic00as :: (HasAgentConfig, HasCallStack) => Blob -> IC00

ic00 :: HasAgentConfig => IC00

ic00as' :: HasAgentConfig => Blob -> Blob -> Text -> Blob -> IO HTTPErrOrReqResponse

callIC :: forall s a b. (HasCallStack, HasAgentConfig) => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => (CandidArg a, CandidArg b) => IC00 -> Blob -> Label s -> a -> IO b

type family UnRec r where ...

Equations

UnRec (Rec r) = r

type PartialSettings r = (Forall r Unconstrained1, (Map Maybe r .// UnRec Settings) ≈ UnRec Settings)

fromPartialSettings :: PartialSettings r => Rec r -> Settings

ic_create :: (HasCallStack, HasAgentConfig, PartialSettings r) => IC00 -> Rec r -> IO Blob

ic_provisional_create :: (HasCallStack, HasAgentConfig, PartialSettings r) => IC00 -> Maybe Natural -> Rec r -> IO Blob

ic_install :: (HasCallStack, HasAgentConfig) => IC00 -> InstallMode -> Blob -> Blob -> Blob -> IO ()

ic_uninstall :: (HasCallStack, HasAgentConfig) => IC00 -> Blob -> IO ()

ic_set_controllers :: HasAgentConfig => IC00 -> Blob -> [Blob] -> IO ()

ic_start_canister :: HasAgentConfig => IC00 -> Blob -> IO ()

ic_stop_canister :: HasAgentConfig => IC00 -> Blob -> IO ()

ic_canister_status :: forall a b. (a -> IO b) ~ (ICManagement IO .! "canister_status") => HasAgentConfig => IC00 -> Blob -> IO b

ic_deposit_cycles :: HasAgentConfig => IC00 -> Blob -> IO ()

ic_top_up :: HasAgentConfig => IC00 -> Blob -> Natural -> IO ()

ic_delete_canister :: HasAgentConfig => IC00 -> Blob -> IO ()

ic_raw_rand :: HasAgentConfig => IC00 -> IO Blob

callIC' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => IC00 -> Blob -> Label s -> a -> IO ReqResponse

ic_create' :: (HasCallStack, HasAgentConfig, PartialSettings r) => IC00 -> Rec r -> IO ReqResponse

ic_provisional_create' :: (HasCallStack, HasAgentConfig, PartialSettings r) => IC00 -> Maybe Natural -> Rec r -> IO ReqResponse

ic_install' :: HasAgentConfig => IC00 -> InstallMode -> Blob -> Blob -> Blob -> IO ReqResponse

ic_update_settings' :: (HasAgentConfig, PartialSettings r) => IC00 -> Blob -> Rec r -> IO ReqResponse

ic_set_controllers' :: HasAgentConfig => IC00 -> Blob -> [Blob] -> IO ReqResponse

ic_delete_canister' :: HasAgentConfig => IC00 -> Blob -> IO ReqResponse

ic_deposit_cycles' :: HasAgentConfig => IC00 -> Blob -> IO ReqResponse

ic_top_up' :: HasAgentConfig => IC00 -> Blob -> Natural -> IO ReqResponse

callIC'' :: forall s a b. HasAgentConfig => KnownSymbol s => (a -> IO b) ~ (ICManagement IO .! s) => CandidArg a => Blob -> Blob -> Label s -> a -> IO HTTPErrOrReqResponse

ic_install'' :: (HasCallStack, HasAgentConfig) => Blob -> InstallMode -> Blob -> Blob -> Blob -> IO HTTPErrOrReqResponse

ic_uninstall'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_set_controllers'' :: HasAgentConfig => Blob -> Blob -> [Blob] -> IO HTTPErrOrReqResponse

ic_start_canister'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_stop_canister'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_canister_status'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_delete_canister'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_deposit_cycles'' :: HasAgentConfig => Blob -> Blob -> IO HTTPErrOrReqResponse

ic_raw_rand'' :: HasAgentConfig => Blob -> IO HTTPErrOrReqResponse

barrier :: HasAgentConfig => [Blob] -> IO ()

enum :: (AllUniqueLabels r, KnownSymbol l, (r .! l) ~ ()) => Label l -> Var r

asHex :: Blob -> String

textual :: Blob -> String

shorten :: Int -> String -> String