IC.Wasm.Winter

Description

This module provides a thin wrapper around the winter Wasm engine, exposing just the bits needed by the IC ref.

This is the interface at which one might plug in a different Wasm engine.

Documentation

type Module = Module Phrase

parseModule :: ByteString -> Either String Module

exportedFunctions :: Module -> [FuncName]

type Import s = (ModName, FuncName, StackType, StackType, [Value] -> HostFunc s)

type Imports s = [Import s]

type HostM s = ExceptT String (ST s)

type HostFunc s = HostM s [Value]

data Value

Constructors

I32 !Int32

I64 !Int64

F32 !Float

F64 !Double

type StackType = [ValueType]

data ValueType

Constructors

I32Type

I64Type

F32Type

F64Type

type Address = Int64

type Size = Int32

getBytes :: Instance s -> Address -> Size -> HostM s ByteString

setBytes :: Instance s -> Address -> ByteString -> HostM s ()

initialize :: forall s. Module -> Imports s -> HostM s (Instance s)

type Instance s = (IntMap (ModuleInst Phrase (ST s)), Int)

invokeExport :: Instance s -> FuncName -> [Value] -> HostM s [Value]

invokeTable :: Instance s -> Int32 -> [Value] -> HostM s [Value]